Slideshow creator functional
SelectImagesController now takes a slider value as input for how many photos to download, and generates a video with a slideshow and text saved to video_files/output.mp4. Ultimately this process will be moved to the create creation process so as to hide the time intensive processing, but I didnt want to mess up any merges you might be attempting. The last step for making this video includes merging the audio files (also knowing the length of the audio files) so thats all I can do on it for now.