Added property binding to eliminate user errors
Used property binding to disable buttons when nothing is entered/selected and to disable buttons in other circumstances. This allows to have a foolproof way to prevent errors on the user's part and additionally, helped remove obtrusive pop-up error messages in several places. 

Also added a feature where the user is presented with a default file name for the file they are saving to help save time on their part and improve the overall user experience.