Fixing the space in search term fault
A fault was occurring when the user entered a search term that had a space in it.
The command was searching to delete any images that began with the search term, but of course ones with a space in them were changed so that they could be a valid file name.
This meant the program was not deleting the originally sized images, leading to duplicates and missized images. 
Now it deletes images with the spaces as well.

While I was at it I also changed the fontsize of the creation so that it would decrease if the length of the search term was too large.