Added loading animation when searching for a term and fixed a few faults
Added a loading animation to the create menu scene. Fixed a fault where the create audio button was crashing the app as the default file naming code was entering an infinite loop. Additionally, the back button on the audio files scene was not working which has now been fixed. Made some minor layout changes to a few scenes. Made the app window non-resizable so that children cannot change its size, which could potentially affect the current layout.