Added an alert for the learn menu when no creations have been created
When there are no existing creations, the user cannot access the creation or the memory game. They are prompted to create a new creation before being able to access the quiz component. 

Also, made minor changes to the wording of the alerts when a voice cannot pronounce a particular word in the selected text.